Cleaning up thread model/lock usage and queueing (thread safety relaxed)
Try to pick a sensible default number of threads based on
hardware concurrency, or OpenMP settings

Tweak the callback execution such that Threadlocal/scalable
endpoints can invoke the callback without putting it onto
a thread local queue - since only the thread doing the polling
can get a completion and the callback can run on that thread anyway.

Tweak the number of completions fetched to be >1 if they are going to
be placed on a queue anyway.

Scalable endpoints require threadcount at init time (scalable
endopoint are not supported properly by hardware so
cannot currently be used, but code might be useful for Tx and
is left in for suture work)

build notes for daint

Revert changes to registration flags

Use new reserved type in request data

Cleanups prior to merge, scalable endpoints not good for ghex usage