Scalable endpoints require threadcount at init time
Try to pick a sensible default number of threads based on
hardware concurrency, or OpenMP settings

Tweak the callback execution such that Threadlocal/scalable
endpoints can invoke the callback without putting it onto
a thread local queue - since only the thread doing the polling
can get a completion and the callback can run on that thread anyway.

Tweak the nukber o completions fetched to be >1 if they are going to
be placed on a queue anyway.